Clean up the giant mess of factories.
End result:

1. Each object (imagestack, intensity table, and codebook) have their own test factory modules.  They produce objects of their respective types.  For example, starfish/codebook/test/factories.py produces `Codebook` objects.
2. Synthetic factories that produce multiple objects live in starfish/test/factories.py
3. Remove a lot of `pytest.fixtures`.  They complicate comprehension and static analysis tools.
4. Remove unused test fixtures.  Believe me, there was a bunch.
5. Got rid of `test_synthetic_spot_creation_produces_an_imagestack`, as that's something that static type analysis should already do.
