+++ starfish/codebook/test/factories.py
@@ -1,0 +1,63 @@
+import tempfile
+
+from starfish.codebook.codebook import Codebook
+from starfish.types import Axes, Features
+
+
+def simple_codebook_array():
+    return [
+        {
+            Features.CODEWORD: [
+                {Axes.ROUND.value: 0, Axes.CH.value: 0, Features.CODE_VALUE: 1},
+                {Axes.ROUND.value: 1, Axes.CH.value: 1, Features.CODE_VALUE: 1}
+            ],
+            Features.TARGET: "SCUBE2"
+        },
+        {
+            Features.CODEWORD: [
+                {Axes.ROUND.value: 0, Axes.CH.value: 1, Features.CODE_VALUE: 1},
+                {Axes.ROUND.value: 1, Axes.CH.value: 1, Features.CODE_VALUE: 1}
+            ],
+            Features.TARGET: "BRCA"
+        },
+        {
+            Features.CODEWORD: [
+                {Axes.ROUND.value: 0, Axes.CH.value: 1, Features.CODE_VALUE: 1},
+                {Axes.ROUND.value: 1, Axes.CH.value: 0, Features.CODE_VALUE: 1}
+            ],
+            Features.TARGET: "ACTB"
+        }
+    ]
+
+
+def loaded_codebook():
+    with tempfile.NamedTemporaryFile() as tf:
+        codebook = Codebook.from_code_array(simple_codebook_array())
+        codebook.to_json(tf.name)
+
+        result = Codebook.from_json(tf.name, n_channel=2, n_round=2)
+    return result
+
+
+def codebook_array_factory() -> Codebook:
+    """
+    Codebook with two codewords describing an experiment with three channels and two imaging rounds.
+    Both codes have two "on" channels.
+    """
+    data = [
+        {
+            Features.CODEWORD: [
+                {Axes.ROUND.value: 0, Axes.CH.value: 0, Features.CODE_VALUE: 1},
+                {Axes.ROUND.value: 1, Axes.CH.value: 1, Features.CODE_VALUE: 1}
+            ],
+            Features.TARGET: "GENE_A"
+        },
+        {
+            Features.CODEWORD: [
+                {Axes.ROUND.value: 0, Axes.CH.value: 2, Features.CODE_VALUE: 1},
+                {Axes.ROUND.value: 1, Axes.CH.value: 1, Features.CODE_VALUE: 1}
+            ],
+            Features.TARGET: "GENE_B"
+        },
+    ]
+    return Codebook.from_code_array(data)
