+++ starfish/intensity_table/test/test_from_imagestack.py
@@ -5,7 +5,28 @@ Tests for IntensityTable.from_image_stack method
 import numpy as np
 
 from starfish import ImageStack, IntensityTable
+from starfish.test.factories import (
+    codebook_intensities_image_for_single_synthetic_spot,
+    synthetic_spot_pass_through_stack,
+)
 from starfish.types import Axes
+
+
+# TODO ambrosejcarr: improve the tests here.
+def test_imagestack_to_intensity_table():
+    codebook, intensity_table, image = codebook_intensities_image_for_single_synthetic_spot()
+    pixel_intensities = IntensityTable.from_image_stack(image)
+    pixel_intensities = codebook.metric_decode(
+        pixel_intensities, max_distance=0, min_intensity=1000, norm_order=2)
+    assert isinstance(pixel_intensities, IntensityTable)
+
+
+def test_imagestack_to_intensity_table_no_noise():
+    codebook, intensity_table, image = synthetic_spot_pass_through_stack()
+    pixel_intensities = IntensityTable.from_image_stack(image)
+    pixel_intensities = codebook.metric_decode(
+        pixel_intensities, max_distance=0, min_intensity=1000, norm_order=2)
+    assert isinstance(pixel_intensities, IntensityTable)
 
 
 # TODO ambrosejcarr: crop is not tested because it should be moved out of this function
